move autoscaler deployment to cluster setup workflow
